Create 3 different recipes into the page
The assignment require me to create a recipes pages so i made one my
favorites dishes in Japan, the Katsudon.

I made the Katsudon entry and the ask Chat gpt to generate two more,
following the structure i constructed with the assignment instructions,
also ask chat gpt to make some images and add them to the project.

Also in this commit i changed the folder structure, not very atomic of
my part.